Created generic html document
this is a generic html file pointing to our javascript. I use this to test the console in my browser.

i am also uploading the initial commit for my javascript. in this file i define our variables, use js functions to loop through an array of characters to use for the password, join 10 results to generated a complex password with 10-characters. I use console.log(generatedPassword) to print the generated password to our console