finalized js, css, and html
created html document and elements that are referenced in the js script. I updated the CSS to add styling that mimics the design in the assignment's readme file. In the js script, I defined variables and constants to store all character category types, and all user prompt inputs. I added a function to generate a random password, and a function to copy the password to your clipboard by pressing a button on the page